+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Fixed - Assert calling cli.title() immediately after creating new option
           group via cli.command(...).
 - Added - Support populating unknownArgs from any command.
+- Fixed - No error when too few values for option named after optional operand,
+          for example: cli.optVec<string>("[FILES] files").size(2, 4);
 
 ## dimcli 7.2.0 (2024-02-22)
 - Added - More cli.to\*Cmdline() and cli.to\*CmdlineL() overloads